feat: add error handling and standardized error response body

+++ src/main/java/io/github/wiltonreis/library/services/AuthorService.java
@@ -1,7 +1,11 @@
 package io.github.wiltonreis.library.services;
 
+import io.github.wiltonreis.library.exception.OperationNotAllowed;
 import io.github.wiltonreis.library.model.Author;
 import io.github.wiltonreis.library.repositories.AuthorRepository;
+import io.github.wiltonreis.library.repositories.BookRepository;
+import io.github.wiltonreis.library.validators.AuthorValidator;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -9,20 +13,21 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Service
+@RequiredArgsConstructor
 public class AuthorService {
 
     private final AuthorRepository authorRepository;
-
-    public AuthorService(AuthorRepository authorRepository) {
-        this.authorRepository = authorRepository;
-    }
+    private final AuthorValidator authorValidator;
+    private final BookRepository bookRepository;
 
     public Author saveAuthor(Author author){
+        authorValidator.validate(author);
         return authorRepository.save(author);
     }
 
     public void updateAuthor(Author author){
         if(author.getId() == null) throw new IllegalArgumentException("Author not registered");
+        authorValidator.validate(author);
         authorRepository.save(author);
     }
 
@@ -31,6 +36,7 @@ public class AuthorService {
     }
 
     public void deleteAuthor(UUID id){
+        if(hasBook(id)) throw new OperationNotAllowed("It is not possible to delete an author who has books.");
         authorRepository.deleteById(id);
     }
 
@@ -44,5 +50,10 @@ public class AuthorService {
 
         return authorRepository.findAll();
     }
+
+    private boolean hasBook(UUID id){
+        return bookRepository.existsByAuthorId(id);
+    }
+
 
 }
